ERR on passing react element
