+++ client/containers/DashboardOverview/overviewRows/PubOverviewRow.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import classNames from 'classnames';
 import { AnchorButton } from '@blueprintjs/core';
 
-import { PubByline } from 'components';
-import { DefinitelyHas, Pub as BasePub } from 'types';
+import { PubByline, Icon } from 'components';
+import { DefinitelyHas, Pub as BasePub, IconLabelPair } from 'types';
 import { pubUrl } from 'utils/canonicalUrls';
 import { getDashUrl } from 'utils/dashboard';
 import { usePageContext } from 'utils/hooks';
@@ -19,6 +19,32 @@ type Props = {
 	className?: string;
 	pub: Pub;
 	inCollection?: boolean;
+};
+
+const defaultLabelPairs = (iconLabelPairs: IconLabelPair[]) => {
+	return (
+		<div className="summary-icons">
+			{iconLabelPairs.map((iconLabelPair, index) => {
+				const { icon, label, iconSize: iconLabelPairIconSize = 12 } = iconLabelPair;
+				const iconElement =
+					typeof icon === 'string' ? (
+						<Icon icon={icon} iconSize={iconLabelPairIconSize} />
+					) : (
+						icon
+					);
+				return (
+					<div
+						className="summary-icon-pair"
+						// eslint-disable-next-line react/no-array-index-key
+						key={index}
+					>
+						{iconElement}
+						{label}
+					</div>
+				);
+			})}
+		</div>
+	);
 };
 
 const PubOverviewRow = (props: Props) => {
@@ -38,16 +64,29 @@ const PubOverviewRow = (props: Props) => {
 			target="_blank"
 		/>
 	);
+
+	// get label pairs
+	// set label pairs in a div
+	// render normal div
+	const labelPairs = defaultLabelPairs([
+		...getScopeSummaryLabels(pub.scopeSummary),
+		getPubReleasedStateLabel(pub),
+	]);
+
+	// if submission get submissinolabel pairs
+	// set them in a div as a variable
+	// render div with submissino status
+
 	return (
 		<OverviewRowSkeleton
 			className={classNames('pub-overview-row-component', className)}
 			href={getDashUrl({ pubSlug: pub.slug })}
 			title={pub.title}
 			byline={<PubByline pubData={pub} linkToUsers={false} truncateAt={8} />}
-			iconLabelPairs={[
+			iconLabelPairs={defaultLabelPairs([
 				...getScopeSummaryLabels(pub.scopeSummary),
 				getPubReleasedStateLabel(pub),
-			]}
+			])}
 			leftIcon={leftIconElement || 'pubDoc'}
 			rightElement={rightElement}
 			darkenRightIcons={inCollection}
