+++ types/iconLabel.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+import { IconName } from 'client/components';
+
+export type IconLabelPair = {
+	icon: IconName;
+	label: string | React.ReactNode;
+	iconSize?: number;
+};
